+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.7.0
 
+* Added timeout
 * Added code coverage to testing
 
 ## v0.6.1