+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## v0.9.3
+
 * Added RngService implementation
 * Added PlatformCrypto implementation